Move all internal errors to LibCnbError
This allows buildpacks to implement their own error handling for these. Buildpacks can format the errors in a way that is consistent with their other error messages and add extra information.
To make things easy for buildpack implementors, a GenericErrorHandler is provided that logs the messages in a generic way, making custom error handling optional.