CI: Remove --all-targets from `cargo {build,test}`
Since unlike `cargo clippy` (where it's required to ensure everything is linted), for
`cargo build` it's somewhat unnecessary (as we don't have any of the additional
targets such as examples or benches (the examples we have aren't detected/built
currently and should probably be their own step anyway), and in the case of 
`cargo test`, actually means less is tested (the doc-tests are omitted?!).

See:
https://doc.rust-lang.org/cargo/commands/cargo-build.html#target-selection
https://doc.rust-lang.org/cargo/commands/cargo-test.html#target-selection

GUS-W-9821318.