Enable all passing clippy::pedantic lints
Clippy has a `pedantic` lint category for lints that are more opinionated,
or that sometimes have a higher false positive rate.

Many of these lints are actually very useful (particularly for those new to Rust,
like we are), and so we should take advantage of them - with the understanding
that it's absolutely fine to disable individual pedantic lints in the future that we
find annoying/counter-productive.

As a first step, this change enables all currently passing pedantic lints, with the
failing lints marked as `allow` until we've had a chance to triage/fix them. This
will prevent additional lint regressions in the time it takes to fix those currently
failing.

The list of all Clippy lints by category can be found here:
https://rust-lang.github.io/rust-clippy/stable/index.html

Note: Lint level `warn` has been chosen, since it doesn't block compilation (and so
temporary lint transgressions during debugging or refactoring won't block compiling),
however will still be caught by CI due to its use of `--deny warnings`. This is the same
approach used for the subjective-but-enabled-by-default lints such as `clippy::style`
and `clippy::complexity`.

Fixes #66.