[close #102] Support layer data under `types` key
In https://github.com/buildpacks/spec/blob/buildpack/0.6/buildpack.md#layer-content-metadata-toml the values for build, cache, and launch were moved under a `types` key in order to make it easier to set the values via bash `echo`.

This change moves the values under a `types` key and adds test documentation showing how to use this with the existing "builder" pattern. We should encourage the builder use since it would make this a non-breaking change for any buildpacks using the builder (rather than directly creating the structs).