Reduce amount of explicit type annotations
By introducing a central type for the buildpack itself, we can leverage associated types to define other types that are valid throughout the buildpack. For example, the type for the buildpack metadata never changes. The same is true about the error type and the target platform. The code has been refactored to reference these associated types instead of requiring explit values everywhere.