Mark `clippy:module_name_repetitions` disablement as permanent
Many of the disabled lints are wanted and just haven't been fixed yet. However,
for this lint, it is too noisy and enforces a style that reduces readability in many
cases. 

As such, it has been moved out of the "to triage/fix" list to indicate it is a long-term
disablement.

Closes #83.