Change DetectResultBuilder and BuildResultBuilder to match LayerResultBuilder (#209)
LayerResultBuilder always wrapped the result of it's build call in a Result to simplify usage when used in a Layer implementation. This change makes the API for DetectResultBuilder and BuildResult builder match, adds documentation why it returns a Result and adds another method to build the respective value without it being wrapped in a Result.