Add remaining unit tests for `libcnb_data::buildpack` (#212)
This adds the last of the missing unit tests for `libcnb_data::buildpack`.

The `Eq` and `PartialEq` Derive attributes had to be added so that the assertions
in the new tests could compare the relevant structs.

Whilst working on this I found #211. Until that is fixed, the meta-buildpack
tests will need to continue to have an invalid `[[stacks]]` table.

GUS-W-10222615.