Remove `test_` prefix from test names
Previously the test names in this repo were inconsistent, with
some tests having the `test_` prefix, and others not.

Since tests are marked with a `#[test]` attribute, the `test_`
prefix seems redundant, so I've chosen to remove it, rather
than add it to the tests that didn't have it.

Fixes #201.