Link to PRs from changelog entries
When consuming other project's changelogs, I find it useful to be able to
look at the PR related to a changelog entry, when the changelog entry is not
clear, or I want to know more about the motivation/exactly what changed.

This adds PR links for anything newer than v0.4.0 (I don't think it's worth the
effort of back-filling previous entries; there was more churn in the project
earlier on + fewer users etc).