Fix the data model and validation of `Buildpack.version`
Previously it was a `semver::Version`, which:
- Allowed for version values not permitted by the CNB spec, such as those
  with a prerelease (eg: `-foo`) or build metadata (eg `+abc`) suffix.
- Had `prerelease` and `build` fields, which are not part of the allowed
  version scheme.

Version parsing/validation is now performed via a new `BuildpackVersion` type,
which follows a similar style to that of `BuildpackApi`. However the implementation
of  `BuildpackVersion::try_from` has had to use regex, since buildpack versions
cannot contain leading zeros (unlike the buildpack API version), which would be
silently accepted by `str::parse()` if we instead used the simpler `.split(".")`approach.

As an added bonus of this fix, we get to drop a dependency :-)

GUS-W-10321651.