Test buildpack cross-compilation & packaging in CI
Adds a new `integration` GitHub action job under the `CI` workflow,
which tests that:
1. `cargo libcnb package` can run successfully with the hello-world buildpack.
2. That the resultant buildpack works when used with `pack build`.

This helps to validate both `libcnb-cargo` and the hello-world example.

The Ruby example buildpack remains without an integration test for now, however
it's something we can test in the future when `libcnb-test` is added, in order to test
both `libcnb-test` itself, and library usage of `libcnb-cargo`,

The `cargo build` step has been removed, since it's redundant given that:
(a) `cargo clippy` is a superset of `cargo check`,
(b) the integration job is cross-compiling anyway (which is more representative).

Lastly, branch builds are skipped for all branches apart from `main`, to prevent
duplicate builds on PRs (given that there would now otherwise be 3x2 builds per PR,
with the new `integration` job).

Fixes #259.
GUS-W-10418196.