Enable caching in CI
Enables caching using:
https://github.com/Swatinem/rust-cache

This action caches the Cargo index cache, as well as dependencies of
libcnb (but nothing else, for cache size efficiency), and automatically
handles setting the cache key based on `**/Cargo.toml`, the Rust
version and the GitHub job name.

This action is much superior to the manual alternatives I had been investigating
prior to finding that action.

Fixes #262.
GUS-W-10419238.