Skip integration tests when using `cargo test` (#316)
The integration tests in the examples and `libcnb-test` are now annotated
with the `#[ignore]` attribute, which causes them to be skipped by default
when using `cargo test`. The skipped tests will be listed in the test log output
as "ignored".

To run the integration tests, `cargo test -- --ignored` must be used.

This has several benefits:
- Allows easily running just the quicker subset of the tests (unit tests
  and doctests) when developing locally, via `cargo test`.
- Means the integration tests can be moved from the CI `test` job to the
  `integration` job (alongside the existing `cargo libcnb package` test),
  which evens out the CI end to end time, and means faster feedback in case
  of failures of the unit tests (particularly since the `test` job no longer needs
  to duplicate all of the musl-related setup steps).

Fixes #301.