Prepare new release of all crates
To pick up the crate metadata updates for the repository move,
as well as a handful of not yet released other changes.

I've added a mention of the repo move to the changelogs to:
(a) raise awareness (eg for updating git remotes)
(b) so it's clearer why the crates with no other changes (aside
      from the metadata) have been released.

GUS-W-10969178.