Switch to `heroku/builder:22` in docs and tests
Since it supersedes `heroku/buildpacks:20`:
https://github.com/heroku/builder#heroku-builder-images

Whilst updating the example build output in the README, I noticed
that the specified commands were incorrect (eg wrong buildpack path),
so those have been fixed.

In addition, I've switched the hello world example to use the wildcard
stack, since the code is not stack-specific, and it will save us having to
remember to update it in the future.

Lastly, updating the Ruby example buildpack to `heroku/builder:22`
revealed that the buildpack was actually broken and relying on
system Ruby, so that buildpack has been excluded from using the
new builder until that is resolved:
https://github.com/heroku/libcnb.rs/issues/398