Fix `libcnb_newtype!` doctest on Rust 1.62+
As of Rust 1.62, doctests that contain non-exported `macro_rules!`
macros are now tested, when they were not before.

This causes `cargo test` to fail for this repository, due to the
`libcnb_newtype!` doctest trying to test a private macro.

Doctests can only test things that are public, so we can to either:
1. Make `libcnb_newtype!` public.
2. Annotate the doctest with `compile_fail`, so failure is expected.
3. Remove the example from the doctest.

I've gone with option (2), since it's useful to have the example (so
that people working on libcnb's codebase can use the macro more
easily), however the macro is really an internal thing, so I don't think
we want to export it publicly.

Fixes #410.
GUS-W-11312457.