Add changelog entries for recent dependency upgrades (#420)
Whilst it's unlikely, there is a chance that dependency upgrades
could cause unforeseen breakage for consumers of libcnb, in
which case a mention of them in the changelog would make it
easier for users to debug.

Also, without a changelog entry, it's easy to not realise that there
are unreleased changes for a package that are waiting on a release
to be performed.

As such, in addition to adding the recent `fancy-regex` and `bollard`
upgrades to the respective changelogs, I've removed the Dependabot
PR exemption from the check-changelog task, to encourage addition
of changelog entries in the future. 

This shouldn't be too much hassle in practice, since this repository
is for a library (and not an end application), so does not commit
its `Cargo.lock`. As such, Dependabot PRs are only for major version
upgrades of top-level dependencies (and not every patch/minor release),
and so are relatively rare.

The `skip changelog` label can still be used in cases where a changelog
entry isn't warranted (for example updates to the GitHub Actions
dependencies, or updates to `[dev-dependencies]`.