Make the Buildpack API version check more robust
Previously, if a `buildpack.toml` either:
1. Didn't match the specific custom buildpack type (used when
  custom fields are specified under `[[metadata]]`)
2. Didn't match the CNB spec

...then a generic "failed to deserialise buildpack.toml" message was
shown, rather than any error about the buildpack API version not
matching.

In cases where the chosen API version is actually the cause of the
error (for example if a future API version makes breaking changes
to the schema), then this means the underlying cause isn't mentioned.

Now, the check uses a minimal `BuildpackDescriptorApiOnly` type
when deserialising `buildpack.toml`, to ensure the API version can
still be determined in those cases.

The `BuildpackDescriptorApiOnly` type was used instead of a raw
`toml::value::Table` since otherwise there are a myriad of `Option`s
and related to handle, from both the potential lack of `api` field and
it not having a value of type string etc.