Remove unnecessary usages of public structs/functions in examples
For the common case of buildpacks being built only as a binary,
and not as a library, it's not necessary to publicly export everything.

When `pub` is used, Clippy cannot warn about deadcode, since it
doesn't know if the struct/function is called via a consumer of the
public API.

Switching to `pub(crate)` avoids this problem.

In addition, the `basics` example has been switched from a private
buildpack struct to `pub(crate)` for the reasons in #290.

Closes #290.