Compile README tests as a part of doctest
In https://github.com/heroku/libcnb.rs/pull/366#issuecomment-1151577355 I left feedback that stated the README was broken. I found we actually DO have doctests on it but even if you add a non-existent method in there it still passes. Why? Because the method has the `#[test]` line on it so it only compiles on the test environment and doctest does not use the test environment.

I removed the config declaration and added a comment explaining why.

In `lib.rs` I moved the ordering to match https://doc.rust-lang.org/rustdoc/write-documentation/documentation-tests.html but I don't think it has an impact.