Disable `fancy-regex` default features
The `fancy-regex` crate (like the `regex` crate, which it wraps), has
a number of additional features enabled by default, that we do not
need and cause the binary size and compilation times to be increased.

Specifically:
- We don't need unicode support in our validation regexes.
- We only use a few simple regexes against small haystack sizes, that
   are only invoked a handful of times (once per field in TOML files) so
   don't need the `regex`'s advanced performance optimisations.

For more information, see:
https://docs.rs/regex/latest/regex/#crate-features

By disabling these, the release+stripped binary size of the `basics`
buildpack (which is a baseline for all other buildpacks, no buildpacks
can be slimmer than this) is reduced from 2.1 MB to 1.3 MB.

Whilst in absolute terms it's not a massive reduction:
- I think it's important for libcnb to be seen as a slim/lightweight
  framework, given people will inevitably compare it to say using
  bash scripts.
- This saving can add up for builders that use a micro-buildpack
  based strategy.
- It increases libcnb.rs' value proposition over some of the alternative
  Go-based CNB frameworks, whose binaries are often larger.

Fixes #365.