Stop using regex to validate `BuildpackVersion` (#435)
The previous implementation used regex due to the spec requirement
of redundant leading zeros not being permitted. However, implementing
this without regex is just as easy, avoids using a heavy dependency (though
sadly we still need it elsewhere in the codebase), closer to the approach used
for `BuildpackApi`, and IMO slightly cleaner.