libcnb-cargo: Disable unused `stderrlog` features
As of `stderrlog` v0.5.3, it's possible to disable the timestamps
feature of `stderrlog` (which we don't use), which otherwise requires
the `chrono` dependency. 

Doing so reduces the number of transitive dependencies by 5, and also
removes Chrono from the dependency tree, which currently has reported
CVEs so can cause security audit false positives.

GUS-W-11378865.