Fix the buildpack size reported by `cargo libcnb package` (#442)
Previously the size of the packaged buildpack reported by
`cargo libcnb package` was incorrect, since `fs_extra::dir::get_size`
doesn't handle symlinks correctly. This meant the reported sizes
counted the buildpack binary size twice, due to the symlink of
`bin/detect` to `bin/build`.

In addition, the size was previously displayed in SI units (and with
only an `M` suffix, not `MB`), when we really want it to use binary
sizes (ie: MiB; divided by `1024*1024` rather than `1000*1000`).

Now, both of these issues are fixed, and the dependency on both the
`fs_extra` (which is unmaintained/not super reliable) + `size_format`
(which is bloated for what we need) crates have been dropped, reducing
the total number of dependencies of `libcnb-cargo` from 62 to 54.

Lastly, if an IO error occurred during calculation of the buildpack size,
this is no longer ignored (by listing the size as "unknown size"), but
instead results in a non-zero exit code. Whilst the size is a nice to
have, calculation of the size should never fail in the common case,
so covering up failures seems more likely to hide things than be an
overall benefit.

Fixes #336.