Strip buildpack binaries when packaging (#445)
As of Rust 1.59, Cargo now supports a `strip` option, which when enabled,
causes the binaries output to stripped of all symbol information:
https://blog.rust-lang.org/2022/02/24/Rust-1.59.0.html#creating-stripped-binaries
https://doc.rust-lang.org/beta/cargo/reference/profiles.html#strip

Doing this significantly reduces the size of the binaries, which:
- Results in smaller overall production builder image sizes, which
  reduces push/pull times.
- Helps speed up workflows where the otherwise large dev/debug builds
  were being used (such as integration tests or local buildpack development),
  since Docker/`pack build`'s transfer of files into the ephemeral builder image
  is very slow.
- Helps reduce any end-user perception of `libcnb.rs` powered CNBs being
  bloated compared to say bash-based CNBs (and also keeps us competitive
  with any Go-based CNB frameworks).

Before/after comparison:
- `examples/basics`:
  - dev build: 22.01 MiB -> 3.26 MiB
  - release build: 5.04 MiB -> 1.29 MiB
- `examples/exec.d`:
  - dev build: 37.40 MiB -> 4.10 MiB
  - release build: 9.02 MiB -> 1.78 MiB
  - `cargo test -- --ignored`: 5.5s -> 2.6s
- `heroku/procfile`:
  - dev build: 43.40 MiB -> 5.16 MiB
  - release build: 6.27 MiB -> 2.17 MiB
  - `cargo test -- --ignored`: 9.1s -> 5.2s
- `heroku/jvm`:
  - dev build: 116.55 MiB -> 15.52 MiB
  - release build: 24.40 MiB -> 7.38 MiB
  - `cargo test --test integration_tests`: 97s -> 82s

Stripping the binaries does not result in any loss of useful backtraces, since
with the cross-compilation and/or usage of MUSL, these did not work even
in debug buildpack builds anyway. Both before and after this change, any
panics still show the originating file/line number, which is all that's generally
required when debugging. Lastly, even if backtraces had previously worked,
they are a pain to enable given that the env has to be set outside of the
buildpack invocation, so can't be used when `clear_env = true` is set etc.

Since Cargo only supports the `strip` option as of Rust 1.59, the minimum
Rust version has been bumped to match.

Fixes #319.
GUS-W-11324847.