Add more integration tests for libcnb-test (#447)
Previously `libcnb-test` itself only had a single integration test, for
one failure case. Whilst some other functionality was tested via the
integration tests under `examples/`, not everything was - and we
don't want to make the examples unrepresentative by cramming
unnecessary tests in just to exercise `libcnb-test`. We also don't
want to test `libcnb-test`'s handling of failure cases in examples.

Now the vast majority of `libcnb-test` is covered by its own tests.
The only exception is `.expose_port()` and related, for which we
can add tests as part of deciding how to handle #280.

Many of these new tests use the `heroku/procfile` buildpack, since
its fast to build (and so the tests fast to run) and requires little in the
way of fixtures.

Fixes #426.
GUS-W-11382673.