Use each crate's README as the root module's rustdocs (#460)
At the moment, there is a bit of a gulf between the longer form
README documentation we have in the GitHub repo, and the
reference documentation shown on `docs.rs` for each crate.

I believe this leads to people not really looking at `docs.rs` as much
when starting out (particularly if they are less familiar with Rust, so
aren't used to checking it by default), making discovery of features
and reference examples harder. For example, as seen in:
https://github.com/heroku/libcnb.rs/pull/366#issuecomment-1151577355
https://github.com/heroku/libcnb.rs/issues/402

By using the crate READMEs as the rustdocs for each crate's root
module, it means the content will show at the top of the crate's
`docs.rs` page - meaning they then show all docs, not just the
reference docs. We can then pitch the `docs.rs` page as the best
entrypoint for documentation, rather than repo READMEs.

As an added bonus, our READMEs will now all be linted for markdown
and code correctness (and not only `libcnb-test`'s README). (This is why
there are a few changes to the repo root README, to fix lint issues.)