libcnb-test: Refactor the API for starting containers (#469)
Previously container configuration was specified using a builder-style API,
which was inconsistent with the API for `TestRunner::run_test`, and also
slightly confusing at first glance (eg: for `TestContext::prepare_container`
what does it mean to "prepare" a container?)

Now, a new `TestContext::start_container` takes a `ContainerConfig`
argument along with the closure, similar to `TestRunner::run_test`
and `TestConfig`.

In addition, a new convenience function for running a shell script (a
common use-case) has been added (`TestContext::run_shell_command`)
which handles more of the boilerplate than 
`PrepareContainerContext::start_with_shell_command` used to. In the
future, this function will also check the exit status of the container (#446).

GUS-W-11415596.