libcnb-test: Overhaul the README and crate docs
Whilst the individual libcnb-test APIs already had rustdocs and
examples, discovering what features existed and where in the
modules/types hierarchy to look for these was hard for users
new to the crate.

Now, the README covers:
- what the test framework does
- prerequisites for the host machine
- examples for all common use-cases
- tips/best practices

Some rustdocs for individual APIs have been updated to match
the style used in the README for consistency.

Lastly, the "experimental" label has been removed, since the
crate is now mostly complete.

GUS-W-11468003.