libcnb-test: Don't set a time window for `logs_now()`
Previously `ContainerContext::logs_now` set the `since` and `until`
options in the logs request sent to the Docker daemon.

Not only are these options unnecessary (since the default is to return
all logs), but in fact setting the values explicitly was resulting in logs
intermittently not being returned in some cases. I believe this may
be due to races between the `SystemTime::now()` call and the time
associated with the log lines in the daemon.

Now, we rely on the default values for `since` and `until`:
https://docs.rs/bollard/latest/bollard/container/struct.LogsOptions.html
https://docs.docker.com/engine/api/v1.41/#tag/Container/operation/ContainerLogs

GUS-W-11451448.