Remove the check versions check
The check was added in #492 as part of the switch to always
releasing all libcnb-related crates with identical version numbers.

At that point, crate versioning was controlled by each crate's `Cargo.toml`,
so it would be easy to forget to update some of the crates for each release
(particularly given anyone familiar with the previous process would be
used to only updating the version of the single crate being released).

However as of #500, the new Cargo workspace shared metadata feature
is being used, such that the single version number in the repo root
`Cargo.toml` is now inherited by any crate using `version.workspace = true`.

This means there is no risk of crate versions getting out of sync,
unless we happen to add a new crate (infrequent occurrence),
don't base it on the template from an existing crate (unlikely, since
that's what we've done every time before), and then also miss the
difference during code review (unlikely, since comparing a new thing
to an existing thing is the first step I take when reviewing such changes,
and I'm sure others do too).

As such, this check is not adding enough value to be worth cluttering
up the CI results overview with yet another check.