Introduce CachedLayer and CachedLayerTypes
The current default for `existing_layer_strategy` will recreate a layer on every subsequent build. Effectively that means that if `cached: true` but `existing_layer_strategy` is not set, then the layer is not cached. This is documented in #371.

This PR introduces a CachedLayer trait that implements `Layer` such that the user MUST define `existing_layer_strategy`. 

Sidenote: Is it possible to assert some code doesn't compile? It would be good to encode this desire into the framework so someone doesn't accidentally provide a default implementation in the future, for example.


Fixes #371