Refactor name input so it has validation
```
$ cargo run --package libcnb-cargo -- libcnb init /tmp/lol --name "lol/"
   Compiling libcnb-cargo v0.11.1 (/Users/rschneeman/Documents/projects/work/libcnb.rs/libcnb-cargo)
    Finished dev [unoptimized + debuginfo] target(s) in 3.63s
     Running `/Users/rschneeman/Documents/projects/work/libcnb.rs/target/debug/cargo-libcnb libcnb init /tmp/lol --name lol/`
error: Invalid value 'lol/' for '--name <NAME_NAMESPACE>': Name after the slash cannot be empty 'lol/'

For more information try '--help'
```

It would be great to also validate that name is a valid BuildpackId, however I couldn't figure out how to include it in this crate. I think I would need to add a dependency on `libcnb_data` and I'm not sure of the implications. If we want this we could also copy the regex, it's pretty small.