Convert "rust learning" quip doc to an mdbook
We reference the "rust learning doc" frequently, for example https://github.com/heroku/libcnb.rs/issues/532#issuecomment-1321608026. It's a good resource and I've long talked about formatting it for public consumption. 

This PR introduces a "rust quickstart" book. The goal is to have a public living document where we can record helpful resources, tools, and quality of life development tips (like IDE settings). The target is the existing development team or new developers coming to Rust for the first time wanting to build a buildpack with libcnb.

One spoken worry is that this will expand until it becomes a bloated or generic tutorial: I tried to outline the vision in the `introduction.md` page and make it clear that it's not a place to link to the blog post you just published. I think it should be a curated resource aimed at brevity and targeted at giving people a guiding star of sorts to learning enough Rust to work with libcnb.
 
On the other end of the spectrum there's a thought that it is simply unneeded since there are other resources. This guide is ours. There are many others like it, but this is ours.

The value is in curation. If it proves to be a drain on maintainer resources, we can always delete it. If it proves to be not useful, we can also remove it.