Refactor `cargo libcnb init` errors
There was a mix of unwrapping and manually calling `exit` and returning Result before. Now we're mostly only returning Result and at the top level manually catching and exiting. 

One downside is that in the `file_collision_prompt` there are several different IO operations that could fail. We could fail to read from a file on disk, or we could fail executing `diff -u` via Command. The result from the OS could be the same error message "no such file or directory" and it won't reference the command we tried to run or the file extension.

I complained about this on twitter the reason is that io::Error cannot depend on std as not every target has a file system. The proposed solution was to use crates to do IO access instead, specifically `camino` and `fs-err`. Thread on mastodon: https://ruby.social/@Schneems/109365338171757170

The upside is that now all functions can call all other functions (except for the top level function) without worrying about accidentally exiting the program. If we call `handle_libcnb_init` in our tests and it fails, it won't just fail the test, it will exit the whole process :(