Switch to the zlib flate2 decompression backend
The `flate2` crate supports multiple backends for performing
decompression:
https://github.com/rust-lang/flate2-rs#backends

The default `miniz_oxide` flate2 backend has poor performance in debug
builds/under QEMU:
https://github.com/rust-lang/flate2-rs/issues/297

Ideally we'd use the fastest `zlib-ng` backend, however it fails to
cross-compile:
https://github.com/rust-lang/libz-sys/issues/93

As such we have to use the next best alternate backend, which is the
`zlib` backend.

(This is the backend that the Python CNB already uses.)

GUS-W-13745779.