Improve the ergonomics of `ContainerContext::address_for_port`
Previously `ContainerContext::address_for_port` would panic for some
failure modes, but return an `Option` for others.

Given that:
- the function is only used in tests
- `libcnb-test` already leans heavily into the "directly panic rather than
   make the tests call unwrap/expect everywhere" approach
- the current `Option` usage is for when there is a bug in the test (ie: the
  test author forgot to call `ContainerConfig::expose_port`)

...then it makes sense to panic instead of returning `None`.

This avoids the `.unwrap()` boilerplate on the callers side, and also simplifies
a future PR to switch from Bollard to using the Docker CLI (this change is being
split out to reduce the size of that PR).

The `expect()` message doesn't conform to the recommended "expect as
precondition" style since the other `expect()`s in the function do not either.
(We can always revisit this as part of the Bollard migration.)