libcnb-test: Fix image cleanup when a pack build unexpectedly succeeds
Previously, if a test with an expected result of `PackResult::Failure`
unexpectedly succeeded, the built app image was left behind since
cleanup was only performed in `TestContext::Drop` (and in this
scenario, a `TestContext` is never created in the first place).

Now the app image is correctly cleaned up.

Fixes #586.
Closes #588.
GUS-W-13903181.