libcnb-test: Prevent integration tests from hanging when offline
Previously a number of `libcnb-test`'s own integration tests used an
intentionally-invalid builder of `libcnb/invalid-builder` in order to
test the handling of various failure modes.

Since the previous builder name parsed as a valid builder name, it
caused `pack build` to make requests to Docker Hub for the non-existent
builder, which meant:
- When offline, it means Pack would hang for 20-30s whilst it attempts
  to retry the builder image pull.
- If someone were ever to publish an actual image at
  `libcnb/invalid-builder` our tests would start to fail, and we'd
  execute whatever that image was.

Now, the tests use a builder name that fails parsing, which still allows
testing of `libcnb-test`'s error handling, but means no requests are
ever made to Docker Hub.

GUS-W-13903458.