Add `find_cargo_workspace` utility method
In preparation for adding support for testing meta-buildpacks the `find_cargo_workspace` method is being added to `libcnb-package` so that it can be shared between `libcnb-package` and `libcnb-test`.

Both the packaging and testing processes will use this to locate buildpacks within a workspace in order to resolve project-level dependencies declared with `libcnb:{buildpack_id}`.