Improved buildpack assembly helpers
In preparation for adding support for testing meta-buildpacks the process of assembling compiled buildpacks is being improved in `libcnb-package` so that they are easier to work with in `libcnb-package` and `libcnb-test`.

Both the packaging and testing processes will use this new functionality to assemble compiled buildpacks into a target location.