Run `cargo upgrade` as part of preparing libcnb releases
We've been previously doing these on an adhoc basis, but it makes
sense to instead just do it as part of the release.

Bumping the in-range versions ensures that consumers of libcnb
get updated transitive dependencies (without having to manually
refresh their own `Cargo.lock`s) that match the dependency
versions tested in CI here (since this repo correctly doesn't use
a lockfile, so will be using these latest in-range versions in CI etc).

Eventually we can move all of these preparing release steps into
a GitHub Actions workflow to reduce the toil of performing releases.