Unpin version of the example alpine builder used in CI
In CI a `pack build` is performed using one of libcnb's example CNBs
and the upstream `cnb/sample-builder:alpine` builder, in order to test
that our static MUSL compile works with non-libc base images.

However, in #594 the version of that builder had to be pinned to
an older release, to work around a temporary incompatibility
between the latest version of the builder and the latest stable
Pack CLI.

That issue has now been resolved:
https://github.com/buildpacks/samples/issues/170

So we can switch back to the latest image version again.